Vision
I started making some mockups of the home page based on what we were
sent  before. It's a PDF so there are comments on each part. The left
side is an example of a shrunken down screen to show how responsiveness
could work. While I was at I just played around with some possible logos
too to save space since the name is so long. But we don't have to use
any of those. The important part is kind of leaning toward a homepage
layout because it will affect how we do the responsiveness within the
sites.